+++ WebContent/minified-dbl-src.js
@@ -2488,7 +2488,7 @@ define('minified', function() {
 
 		if (element) {
 			if (isString(spec)) {
-				var name = replace(replace(spec, /^%/, 'data-'), /^[$@]/);
+				var name = replace(replace(spec, /^%/, 'data-'), /^[$@]+/);
 				var s;
 				if (spec == '$')
 					s = element.className;
@@ -2501,7 +2501,7 @@ define('minified', function() {
 						s = element.getAttribute('style');
 				}
 				// @condblock fadeslide
-				else if (/\$\$/.test(spec) && (element['style']['visibility'] == 'hidden' || element['style']['display'] == 'none')) {
+				else if (/^\$\$/.test(spec) && (element['style']['visibility'] == 'hidden' || element['style']['display'] == 'none')) {
 					s = 0;
 				}
 				else if (spec == '$$fade') {
@@ -2516,7 +2516,7 @@ define('minified', function() {
 					s = self['get']('$height');
 				}
 				// @condend fadeslide
-				else if (/^\$/.test(spec)) {
+				else if (/^\$[^$]/.test(spec)) {
 					// @condblock ie8compatibility 
 					if (!_window.getComputedStyle)
 						s = (element.currentStyle||element['style'])[name];
@@ -2578,8 +2578,11 @@ define('minified', function() {
 	 *              between 0 and 1. '$$slide' will also automatically control the element's 'visibility' and 'display' styles. If the value is 0,
 	 *             the element's visibility will automatically be set to 'hidden'. If the value is larger, the visibility will be set to 
 	 *             'visible' and the display style to 'block'. '$$slide' only works with block elements.</td></tr>
+	 * <tr><td>$$scrollX, $$scrollY</td><td>$$scrollY</td><td>Scroll Coordinates</td><td>The names '$$scrollX' and
+	 *             '$$scrollY' can be used on <code>$(window)</code> to set the scroll coordinates of the document.
+	 *             The coordinates are specified in pixels.</td></tr>
 	 * </table>
-	 * 
+	 *  (use on <code>$(window)</code>)
 	 * @example Unchecking checkboxes:
 	 * <pre>
 	 * $('input.checkbox').set('checked', false);
@@ -2655,7 +2658,8 @@ define('minified', function() {
 	 *             '%' also is used to set attributes, but automatically adds 'data-' to the name. 
 	 *             A dollar ('$') prefix is a shortcut for CSS styles. A simple dollar ('$') as name modifies CSS classes.
 	 *             The special name '$$' allows you to set the <var>style</var> attribute in a browser independent way.
-	 *             The special name '$$fade' and '$$slide' create fade and slide effects, and both expect a value between 0 and 1. 
+	 *             The special names '$$fade' and '$$slide' create fade and slide effects, and both expect a value between 0 and 1. 
+	 *             The special names '$$scrollX' and '$$scrollY' allow you to specify the scroll position (use on <code>$(window)</code>). 
 	 *             
 	 * 
 	 * @param value the value to set. If value is null and name specified an attribute, the attribute will be removed.
@@ -2703,7 +2707,7 @@ define('minified', function() {
     		 else
     			// @condend fadeslide
     			 flexiEach(self, function(obj, c) {
-    				 var nameClean = replace(replace(name, /^%/,'data-'), /^[@$]/);
+    				 var nameClean = replace(replace(name, /^%/,'data-'), /^[@$]+/);
     				 var className = obj['className'] || '';
     				 var newObj = /^\$/.test(name) ? obj.style : obj;
     				 var newValue = isFunction(value) ? value($(obj).get(name), c, obj) : value;
@@ -2727,6 +2731,10 @@ define('minified', function() {
 						// @condend
 							setAttr(obj, 'style', newValue);
 					 }
+   				 	 else if (name == '$$scrollX')
+			 			 _window['scroll'](newValue, _window['scrollY']);
+   				 	 else if (name == '$$scrollY')
+			 			 _window['scroll'](_window['scrollX'], newValue);
     				 else if (!/^[@%]/.test(name))
     					 newObj[nameClean] = newValue;
     				 else
@@ -3353,8 +3361,7 @@ define('minified', function() {
 	 * When you animate colors, <var>animate()</var> is able to convert between the three notations rgb(r,g,b), #rrggbb or #rgb. You can use them interchangeably, but you can not 
 	 * use color names such as 'red'.
 	 *
-	 * You can prefix any number, including numbers with units, with "-=" or "+=" in order to specify a value relative to the starting value. The new value will be added
-	 * to or substracted from the start value to determine the end value.
+	 * Instead of the end value, you can also specify a <code>function(oldValue, index, obj)</code> to calculate the actual end value. 
 	 *
 	 * To allow more complex animation, <var>animate()</var> returns a ##promise#Promise## that is fulfulled when the animation has finished. 
 	 *
@@ -3364,16 +3371,15 @@ define('minified', function() {
 	 *                  .animate({$left: '50px', $top: '100px'}, 1000);  // animation
 	 * </pre>
 	 *
-	 * @example Using relative values for animation:
-	 * <pre>
-	 * $('#myMovingDiv').set({$left: '100px', $top: '100px'})                // start values
-	 *                  .animate({$left: '-=50px', $top: '+=100px'}, 1000);  // animation
-	 * </pre>
-	 * 
 	 * @example Change the color of an element:
 	 * <pre>
 	 * $('#myBlushingDiv').set({$backgroundColor: '#000000'})
 	 *                    .animate({$backgroundColor: '#ff0000'}, 1000);
+	 * </pre>
+	 * 
+	 * @example Using a function to calulate the values for animation:
+	 * <pre>
+	 * $('#myMovingDiv').animate({$left: function(v) { return (parseFloat(v) + 100) + 'px'; }}, 1000);  
 	 * </pre>
 	 * 
 	 * @example Fade-out effect:
@@ -3402,8 +3408,16 @@ define('minified', function() {
 	 *
 	 *
 	 * @param properties a property map describing the end values of the corresponding properties. The names can use the
-	 *                   set() syntax ('@' prefix for attributes, '$' for styles, '$$fade' for fading and '$$slide' for slide effects). 
-	 *                   Values must be either numbers, numbers with units (e.g. "2 px") or colors ('rgb(r,g,b)', '#rrggbb' or '#rgb'). 
+	 *                   set() syntax ('@' prefix for attributes, '$' for styles, '$$fade' for fading,  '$$slide' for slide effects, 
+	 *                   '$$scrollX' and '$$scrollY' for scroll coordinates). 
+	 *                   Values must be either numbers, numbers with units (e.g. "2 px") or colors ('rgb(r,g,b)', '#rrggbb' or '#rgb') or
+	 *                   a <code>function(oldValue, index, obj)</code>  to determine the new value. The function  will be invoked for each list element 
+	 *                   to evaluate the new value: 
+	 *                   <dl><dt>oldValue</dt><dd>The old value of the property to be changed, as returned by ##get().
+	 *                   For the CSS style names, this is the computed style of the property </dd>
+	 *                   <dt>index</dt><dd>The list index of the object owning the property</dd>
+	 *                   <dt>obj</dt><dd>The list element owning the property.<dd>
+	 *                   <dt class="returnValue">(callback return value)</dt><dd>The destination value to be set.</dd></dl> 
 	 *                   Number values, including those with units, can be prefixed with "+=" or "-=", meaning that the value is relative 
 	 *                   to the original value and should be added or subtracted.
 	 * @param durationMs optional the duration of the animation in milliseconds. Default: 500ms.
@@ -3434,7 +3448,6 @@ define('minified', function() {
 		// @cond debug var colorRegexp = /^(rgb\(\s*\d+\s*,\s*\d+\s*,\s*\d+\s*\)|#\w{3}|#\w{6})\s*$/i;
 		var self = this;
 		var initState = []; // for each item contains a map {s:{}, e:{}, o} s/e are property name -> startValue of start/end. The item is in o.
-		var numRegExp = /-?[\d.]+/;
 		var loopStop;
 		var prom = promise();
 		var interpolate = isFunction(linearity) ? linearity : function(startValue, endValue, t) {
@@ -3446,17 +3459,12 @@ define('minified', function() {
 		durationMs = durationMs || 500;
 		linearity = linearity || 0;
 		
-		
 		// find start values
-		flexiEach(self, function(li) {
+		flexiEach(self, function(li, index) {
 			var p = {o:$(li), e:{}}; 
 			eachObj(p.s = p.o.get(properties), function(name, start) {
-				var dest = properties[name];
-				if (name == '$$slide') 
-					dest = dest*getNaturalHeight(p.o) + 'px';
-				p.e[name] = /^[+-]=/.test(dest) ?
-						replace(dest.substr(2), numRegExp, extractNumber(start) + extractNumber(replace(dest, /\+?=/))) 
-						: dest;
+				p.e[name] = isFunction(dest) ? dest(start, index, li) : 
+					name == '$$slide' ? properties[name]*getNaturalHeight(p.o) + 'px' : properties[name];
 			});
 			initState.push(p);
 		});
@@ -3489,7 +3497,7 @@ define('minified', function() {
 								newValue += Math.round(interpolate(getColorComponent(start, i), getColorComponent(end, i), t)) + (i < 2 ? ',' : ')');
 						}
 						else 
-							newValue = replace(end, numRegExp, toString(interpolate(extractNumber(start), extractNumber(end), t)));
+							newValue = replace(end, /-?[\d.]+/, toString(interpolate(extractNumber(start), extractNumber(end), t)));
 						isi.o.set(name, newValue);
 					});
 				});